Refactor files
Split and move files to folder acording to it's functionality. Ehance Readability and Maintainability
