+++ src/apps/api/manajemen-piket/handlers.ts
@@ -6,7 +6,7 @@ import {
   readAllMembers,
 } from "./manajemen-piket";
 import { StatusCodes } from "http-status-codes";
-import { logger } from "../../utils/logger";
+import { logger } from "../../../utils/logger";
 
 export const addMemberHandler = async (req: Request, res: Response) => {
   const { memberName } = req.body;
