--- src/apps/lihat_daftar_piket/types.ts
@@ -1,9 +1,0 @@
-export interface Piket{
-    nama: string;
-    tempat: string;
-}
-
-export interface LihatDaftarPiket{
-    piketData: Piket[];
-    tanggalPiket: string;
-}
