+++ src/config/app-config.ts
@@ -1,0 +1,24 @@
+import express from 'express';
+import path from 'path';
+import dotenv from 'dotenv';
+import helmet from 'helmet';
+import cors from 'cors';
+
+dotenv.config();
+
+const app = express();
+
+// Middleware setup
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+app.use(helmet()); // Set security headers
+app.use(cors());
+
+// Serve static files
+app.use(express.static(path.join(__dirname, "../../public")));
+
+// Set view engine to EJS
+app.set("view engine", "ejs");
+app.set("views", path.join(__dirname, "../../views"));
+
+export default app;
