+++ src/middlewares/error-handler.ts
@@ -1,0 +1,12 @@
+import { Request, Response, NextFunction } from "express";
+import { logger } from "../utils/logger";
+
+export const errorHandler = (
+  err: Error,
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
+  res.status(500).send("Something broke!");
+  logger.error(err.stack);
+};
