+++ src/server.ts
@@ -1,65 +1,9 @@
-import express, { Request, Response, NextFunction } from "express";
-import path from "path";
-import dotenv from "dotenv";
-import helmet from "helmet";
-import cors from "cors";
-import { logger } from "./utils/logger";
-import { LihatDaftarPiket } from "./apps/lihat_daftar_piket/types";
-import manajemenPiketRoutes from "./apps/manajemen-piket/routes";
-import daftarAdminRoute from "./apps/daftar-admin/routes";
-import loginRoute from "./apps/login/routes";
+import app from './config/app-config';
+import { errorHandler } from './middlewares/error-handler';
+import routes from './routes/render';
 
-dotenv.config();
-const app = express();
-
-// Middleware setup
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-app.use(helmet()); // Set security headers
-app.use(cors());
-
-// Serve static files
-app.use(express.static(path.join(__dirname, "../public")));
-
-// register routes
-const api_ver = process.env.API_VER || "v1";
-app.use(api_ver, manajemenPiketRoutes);
-app.use(api_ver, daftarAdminRoute);
-app.use(api_ver, loginRoute);
-
-// Set view engine to EJS
-app.set("view engine", "ejs");
-app.set("views", path.join(__dirname, "../views"));
-
-// Sample data
-const piketData = [
-  { nama: "Sian", tempat: "Dapur" },
-  { nama: "Arbi", tempat: "Kamar Mandi" },
-  { nama: "Yoga", tempat: "Tamu" },
-  { nama: "Agus", tempat: "Dapur" },
-];
-
-const tanggalPiket = "Minggu, 17 Juli 2024";
-
-// Routes setup
-app.get("/", (req: Request, res: Response) => {
-  const data: LihatDaftarPiket = { piketData, tanggalPiket };
-  res.render("home", data);
-});
-
-app.use((req: Request, res: Response, next: NextFunction) => {
-  console.log(`Request URL: ${req.url}`);
-  next();
-});
-
-// Error handling middleware
-app.use((err: Error, req: Request, res: Response, next: NextFunction) => {
-  console.error(err.stack);
-  res.status(500).send("Something broke!");
-  logger.error(err.stack);
-});
-
-
+app.use(routes);
+app.use(errorHandler);
 
 // Start the server
 const PORT = process.env.PORT || 3000;
